requirements.txt
aioEasyPillow==0.0.3
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
packaging==23.1
Pillow==9.1.1
pluggy==1.0.0
pytest==7.3.1
typing_extensions==4.5.0
yarl==1.9.2
